Rename `@bento/core/bech32` -> `@bento/core/address`
